# gh-128118: Improve performance of copy.copy by using a fast lookup for atomic and container types