--- a/Misc/NEWS.d/next/Library/2024-12-20-10-57-10.gh-issue-128118.mYak8i.rst
+++ b/Misc/NEWS.d/next/Library/2024-12-20-10-57-10.gh-issue-128118.mYak8i.rst
@@ -1,0 +1,1 @@
+Improve performance of :func:`copy.copy` by adding a fast path for atomic types and container types.